requirements.txt
streamlit==1.29.0
firebase-admin==6.3.0
python-dotenv==1.0.0
qdrant-client==1.7.0
sentence-transformers==2.2.2
PyPDF2==3.0.1
langchain==0.1.0
openai==1.6.1
numpy==1.26.3
python-dateutil==2.8.2